# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "amanvermame786@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "http://my-default-host.com",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "api/register": {
            "post": {
                "summary": "register",
                "description": "register the user for login",
                "parameters": [
                    {
                        "name": "{firstname}",
                        "in": "path",
                        "description": "FirstName",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "{lastname}",
                        "in": "path",
                        "description": "LastName",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "{email}",
                        "in": "path",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "{password}",
                        "in": "path",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "{C_password}",
                        "in": "path",
                        "description": "Confirm Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful Register"
                    },
                    "201": {
                        "description": "error with user"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "api/login": {
            "post": {
                "summary": "login",
                "description": "login the user which email & password matched",
                "parameters": [
                    {
                        "name": "{email}",
                        "in": "path",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "{password}",
                        "in": "path",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful Login"
                    },
                    "204": {
                        "description": "unauthoried user"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
    "definitions": {}
}